change uint16_t to uint32_t to allow more than 32000 PRGs

+++ include/pangenome/pangraph.h
@@ -57,7 +57,8 @@ public:
 
     void remove_low_covg_nodes(const uint32_t &);
 
-    void split_node_by_reads(unordered_set<ReadPtr> &, vector<uint16_t> &, const vector<bool> &, const uint16_t);
+    void split_node_by_reads(unordered_set<ReadPtr> &, vector<uint_least32_t> &, const vector<bool> &,
+                             const uint_least32_t);
 
     //unordered_set<ReadPtr> find_reads_on_node_path(const std::vector<uint16_t>, const std::vector<bool> );
     void add_hits_to_kmergraphs(const std::vector<LocalPRG *> &);
